optimize and rebuild
