+++ app.js
@@ -1,4 +1,4 @@
-import './styles/home.css';
+import './styles/main.css';
 
 import React from 'react';
 import { render } from 'react-dom';
