+++ components/content/content.scss
@@ -7,6 +7,8 @@
     margin-top: 50px;
     height: 300px;
     background-color: #eee;
+    // box-shadow: 50px 50px 100px rgba(0, 0, 0, .6) inset;
+    box-shadow: 10px 10px 80px #666 inset, -10px -10px 80px #666 inset;
     text-align: center;
     line-height: 300px;
     font-size: 50px;
