+++ components/content/index.js
@@ -6,11 +6,7 @@ export default class Content extends Component {
   render() {
     return (
       <div className="content">
-        <div className="placeholder">Content</div>
-        <div className="placeholder">Content</div>
-        <div className="placeholder">Content</div>
-        <div className="placeholder">Content</div>
-        <div className="placeholder">Content</div>
+        {this.props.children}
       </div>
     );
   }
