--- components/top-banner/style/index.js
@@ -1,1 +1,0 @@
-import './index.scss';
