+++ components/top-banner/top-banner.scss
@@ -9,5 +9,5 @@
 
   text-align: center;
   line-height: 400px;
-  font-size: 50px;
+  font-size: 80px;
 }
