+++ index.html
@@ -2,8 +2,8 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>React Webpack Web APP Demo</title>
-  <link rel="stylesheet" href="/build/home.css">
+  <title>React Webpack SPA</title>
+  <link rel="stylesheet" href="/build/main.css">
 </head>
 <body>
   <div id="app"></div>
