+++ pages/login/index.js
@@ -2,12 +2,12 @@ import React, { Component } from 'react';
 import Content from '../../components/content';
 import Footer from '../../components/footer';
 
-class Page1 extends Component {
+class Login extends Component {
   render() {
     return (
       <div>
         <Content>
-          <div className="page1 placeholder">Page1</div>
+          <div className="login placeholder">Login</div>
         </Content>
         <Footer/>
       </div>
@@ -15,4 +15,4 @@ class Page1 extends Component {
   }
 }
 
-export default Page1;
+export default Login;
