+++ pages/login/login.scss
@@ -1,0 +1,8 @@
+.content {
+  .login.placeholder {
+    margin-left: auto;
+    margin-right: auto;
+    width: 600px;
+    background-color: #666;
+  }
+}
