+++ pages/page1/page1.scss
@@ -1,0 +1,6 @@
+.content {
+  .page1.placeholder {
+    background-color: #999;
+    color: #fff;
+  }
+}
