+++ pages/page2/index.js
@@ -1,12 +1,14 @@
 import React, { Component } from 'react';
 import Content from '../../components/content';
-import Footer from '../../components/footer/Footer';
+import Footer from '../../components/footer';
 
 class Page2 extends Component {
   render() {
     return (
       <div>
-        <Content></Content>
+        <Content>
+          <div className="page2 placeholder">Page2</div>
+        </Content>
         <Footer/>
       </div>
     );
