+++ pages/page2/page2.scss
@@ -1,0 +1,5 @@
+.content {
+  .page2.placeholder {
+    background-color: #666;
+  }
+}
