+++ styles/common.scss
@@ -1,3 +1,4 @@
+// common
 body {
   color: #666;
 }
@@ -11,4 +12,10 @@ a,
 a:visited,
 a:active {
   color: #36f;
+}
+
+// common layout
+.content,
+.site-footer {
+  margin-top: 50px;
 }
