--- styles/home.css
@@ -1,11 +1,0 @@
-@import './reset.scss';
-@import './common.scss';
-@import '../components/header/header.scss';
-@import '../components/top-banner/style/index.scss';
-@import '../components/content/style/index.scss';
-@import '../components/footer/footer.scss';
-
-.content,
-.site-footer {
-  margin-top: 50px;
-}
