--- styles/home.scss
@@ -1,11 +1,0 @@
-@import './reset';
-@import './common';
-@import '../components/header/header';
-@import '../components/top-banner/style/index';
-@import '../components/content/style/index';
-@import '../components/footer/footer';
-
-.content,
-.site-footer {
-  margin-top: 50px;
-}
