+++ styles/main.css
@@ -1,0 +1,16 @@
+/* base */
+@import './reset.scss';
+@import './common.scss';
+
+/* common modules*/
+@import '../components/header/header.scss';
+@import '../components/top-banner/top-banner.scss';
+@import '../components/content/content.scss';
+@import '../components/footer/footer.scss';
+
+
+/* pages */
+/* @import '../pages/home/home'; */
+@import '../pages/page1/page1.scss';
+@import '../pages/page2/page2.scss';
+@import '../pages/login/login.scss';
