+++ webpack.production.config.js
@@ -8,7 +8,7 @@ var ExtractTextPlugin = require('extract-text-webpack-plugin');
 
 module.exports = {
   entry: [
-    './scripts/main.js'
+    './app.js'
   ],
   output: {
     path: __dirname + '/build',
@@ -29,7 +29,7 @@ module.exports = {
     ]
   },
   plugins: [
-    new ExtractTextPlugin('home.css'),
+    new ExtractTextPlugin('main.css'),
     new webpack.optimize.OccurrenceOrderPlugin(),
     new webpack.optimize.UglifyJsPlugin({
       compressor: {
